+++ student-course/src/main/java/org/example/studentcourse/service/StudentService.java
@@ -3,6 +3,8 @@ package org.example.studentcourse.service;
 import org.example.studentcourse.dto.request.StudentRequestDto;
 import org.example.studentcourse.entity.Course;
 import org.example.studentcourse.entity.Student;
+import org.example.studentcourse.exception.CourseNotFoundException;
+import org.example.studentcourse.exception.StudentNotFoundException;
 import org.example.studentcourse.repository.CourseRepository;
 import org.example.studentcourse.repository.StudentRepository;
 import org.example.studentcourse.dto.response.StudentCourseResponse;
@@ -26,7 +28,7 @@ public class StudentService
     public StudentResponse addStudent(StudentRequestDto dto)
     {
         Course course = courseRepository.findById(dto.courseId())
-                .orElseThrow(() -> new RuntimeException("Course not found"));
+                .orElseThrow(() -> new CourseNotFoundException("Course not found with id: " + dto.courseId()));
 
         Student student = new Student();
         student.setName(dto.name());
@@ -46,14 +48,14 @@ public class StudentService
     public StudentResponse getStudentById(int id)
     {
         Student student = studentRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Student not found"));
+                .orElseThrow(() -> new StudentNotFoundException("Student not found with id: " + id));
         return mapToResponse(student);
     }
 
     public StudentCourseResponse getStudentsByCourse(int courseId)
     {
         Course course = courseRepository.findById(courseId)
-                .orElseThrow(() -> new RuntimeException("Course not found"));
+                .orElseThrow(() -> new CourseNotFoundException("Course not found with id: " + courseId));
 
         List<StudentResponse> students = studentRepository.findByCourse(course).stream()
                         .map(this::mapToResponse)
@@ -64,6 +66,11 @@ public class StudentService
 
     public void deleteStudent(int id)
     {
+        if (!studentRepository.existsById(id))
+        {
+            throw new StudentNotFoundException("Student not found with id: " + id);
+        }
+
         studentRepository.deleteById(id);
     }
